final_project_python_for_de/pyproject.toml
[project]
name = "final-de-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "pandas>=3.0.2",
    "pnumpy>=2.0.23",
    "requests>=2.33.1",
]
